manifestを簡素化（Actionレベルでクラス化）
