+++ Makefile
@@ -2,7 +2,7 @@
 srcDir      := src
 outDir      := dist
 assetDir    := assets
-entryPoints := $(shell find $(srcDir) -type f -name \*-script*)
+entryPoints := $(srcDir)/content-script.ts $(srcDir)/background-script.ts
 
 srcFiles      := $(shell find $(srcDir) -type f) 
 assertFiles   := $(shell find $(assetDir) -type f)
