+++ src/content-script-action.ts
@@ -1,0 +1,5 @@
+
+export interface ContentScriptAction {
+    setup(): void;
+    excute(): void;
+}
