+++ src/message-type.ts
@@ -3,5 +3,4 @@ export enum MessageType {
     GetGenreWordConversionMap,
     GetGenerWordConversionMode,
     ReplaceGenreWord,
-    EchoMessage,
 }
